Translate project to English

+++ src/my_lib/include/structures.h
@@ -3,7 +3,7 @@
 
 struct RelativeIndex {
   size_t doc_id;
-  float rank;
+  double rank;
 
   bool operator==(const RelativeIndex& other) const {
     return (doc_id == other.doc_id && rank == other.rank);
